% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age2int}
\alias{age2int}
\title{Infer age class intervals from lower age bounds.}
\usage{
age2int(Age, OAG = TRUE, OAvalue = NA)
}
\arguments{
\item{Age}{integer or numeric. Vector of lower age group bounds.}

\item{OAG}{logical. Whether or not the final age group is open.
Default \code{TRUE}.}

\item{OAvalue}{numeric or integer. The value to use for the final age
interval if \code{OAG = TRUE}. Default \code{NA}.}
}
\value{
Age interval vector, of same length as \code{Age}.
}
\description{
Determine age class intervals based on a vector of age class lower bounds.
}
\details{
If the final age group is open, it is given a value of \code{NA} by default,
or else a user-determined value. If the final age group is closed, it is
assumed to be equal to the next-lower interval. If the final age interval is
known and not equal to the next lowest interval, specify \code{OAG = TRUE}
and assign its value to \code{OAvalue}.
}
