% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcAgeN}
\alias{calcAgeN}
\title{Calculate which large age group single ages belong to.}
\usage{
calcAgeN(Age, N = 5, shiftdown = 0)
}
\arguments{
\item{Age}{integer. Vector of single ages (lower bound).}

\item{N}{integer. Desired width of resulting age groups.}

\item{shiftdown}{integer. Move the grouping down by one or more single ages. Optional argument.}
}
\value{
An integer vector of \code{length(Age)} indicating the age group that
each single age belongs to.
}
\description{
Assign single ages to age groups of equal and arbitrary width, and also
optionally shifted.
}
\details{
If you shift the groupings, then the first age groups may have a negative
lower bound (for example of -5). These counts would be discarded for the
oscillatory version of Sprague smoothing, for example, but they are preserved
in this function. The important thing to know is that if you shift the
groups, the first and last groups won't be N years wide. For example if
\code{shiftdown} is 1, the first age group is 4-ages wide.
}
