% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate.R
\name{graduate}
\alias{graduate}
\title{Graduate grouped data}
\usage{
graduate(
  Value,
  Age,
  AgeInt = age2int(Age),
  OAG = TRUE,
  OAnew = max(Age),
  method = c("sprague", "uniform"),
  keep0 = FALSE,
  constrain = FALSE,
  ...
)
}
\arguments{
\item{Value}{numeric vector, presumably counts in grouped ages}

\item{Age}{integer vector, lower bounds of age groups}

\item{AgeInt}{integer vector, age interval widths}

\item{OAG}{logical, default = \code{TRUE} is the final age group open?}

\item{OAnew}{integer, optional new open age, higher than \code{max(Age)}. See
details.}

\item{method}{character, either \code{"sprague"} or \code{"uniform")}}

\item{keep0}{logical. Default \code{FALSE}. If available, should the value in
the infant age group be maintained, and ages 1-4 constrained?}

\item{constrain}{logical. Default \code{FALSE}. Should output be constrained
to sum within the input age groups?}

\item{...}{extra arguments}
}
\description{
A wrapper function for \code{"sprague"} and \code{"uniform"} graduation
methods
}
\seealso{
\code{\link{graduate_sprague}}, \code{\link{graduate_uniform}}
}
