% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interp}
\alias{interp}
\title{Interpolate between two population age distributions.}
\usage{
interp(
  popmat,
  datesIn,
  datesOut,
  method = c("linear", "exponential", "power"),
  power = 2,
  extrap = FALSE,
  negatives = FALSE,
  ...
)
}
\arguments{
\item{popmat}{numeric. An age-period matrix of data to interpolate over. Age
in rows, time in columns.}

\item{datesIn}{vector of dates. The exact dates in each column. See details
for ways to express it.}

\item{datesOut}{vector of dates. The desired dates to interpolate to. See
details for ways to express it.}

\item{method}{string. The method to use for the interpolation, either
\code{"linear"}, \code{"exponential"}, or \code{"power"}. Default
\code{"linear"}.}

\item{power}{numeric power to interpolate by, if \code{method = "power"}.
Default 2.}

\item{extrap}{logical. In case \code{datesOut} is out of range of \code{datesIn},
do extrapolation using slope in extreme pairwise. Default \code{FALSE}.}

\item{negatives}{logical. In case negative output are accepted, set to
\code{TRUE}. Default \code{FALSE}.}

\item{...}{arguments passed to \code{stats::approx}. For example,
\code{rule}, which controls extrapolation behavior.}
}
\value{
numeric matrix (age-period) (or vector if \code{length(datesOut) ==
  1} of the interpolated data for the requested dates.
}
\description{
The interpolation is done by age using a linear, exponential, or power
function. This comes from the PAS spreadsheet called \code{AGEINT}. Be aware
that this is not cohort-component interpolation.
}
\details{
The age group structure of the output is the same as that of the input.
Ideally, \code{datesOut} should be within the range of \code{datesIn}. If
not, the left-side and right-side output are held constant outside the range
if \code{rule = 2} is passed in, otherwise \code{NA} is returned (see
examples). Dates can be given in three ways 1) a \code{Date} class object, 2)
an unambiguous character string in the format \code{"YYYY-MM-DD"}, or 3) as a
decimal date consisting in the year plus the fraction of the year passed as
of the given date.

For methods \code{"exponential"} and \code{"power"}, calculations are carried
out using linear interpolation through \code{log(popmat)}, or
\code{popmat^(1/power)} respectively, then back-transformed. If the data
contain 0s, \code{method = "exponential"} will fail, but \code{"power"} would
still generally work.
}
